requirements.txt
# Top-level dependencies only. Transitive deps are resolved at install time.
discord-webhook
python-dotenv
pyyaml
openai
uagents
pytest
